backend/requirements.txt
django>=5.0,<5.2
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<4.10
psycopg[binary]>=3.3,<3.4
python-dotenv>=1.2,<1.3
gunicorn>=22.0,<23.0
